api_gateway_stack/api_gateway_stack/authorizer_lambda/requirements.txt
urllib3<2
pyjwt
pyjwt[crypto]